% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendist_parse.R
\name{gendist_parse}
\alias{gendist_parse}
\title{Parse Amino Acid Substitutions Per Genome Segment}
\usage{
gendist_parse(test, focal = NULL)
}
\arguments{
\item{test}{a dataframe containing nextclade metadata output for all
sequences in the test set (often circulating lineages within a certain
location and timeframe). Function works on the aaSubstitutions column.}

\item{focal}{optional parameter that takes a dataframe containing nextclade
metadata output for a single focal sequence.}
}
\value{
if a focal parameter is supplied, the function returns a dataframe
with 37 columns for each row in test. This include all aasubs per segment,
all aasubs unique to test per segment, and all aasubs unique to focal per
segment. If no focal is supplied, a dataframe with 13 columns for each row
in test is returned. This corresponds parses all aasubs per sequence
without comparison to a focal lineage.
}
\description{
#' \code{gendist_parse()} calculates parse the aaSubs using the "aaSubstitutions"
column generated from a Nextclade metadata output
(\url{https://clades.nextstrain.org/}). Output includes overall, unique_test, and
unique_focal.
}
\examples{

# Create test df-----------------------------------------------------------
test <- data.frame(
"aaSubstitutions" = c("E:XXX,E:XX1,S:XXX", "E:XXX,E:XX2,M:XXX"),
"seqName" = c(1, 2)
)

# Create focal df----------------------------------------------------------
focal <- data.frame(
"aaSubstitutions" = c("E:XXX,E:XX1,E:XX3,S:XXX,MXX1"),
"seqName" = c("3")
)

# Execute function---------------------------------------------------------
results <- gendist_parse(test, focal)
results

results_with_focal <- gendist_parse(test, focal)
results_with_focal

}
